requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==4.2.9
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.3
pillow==10.2.0
proglog==0.1.10
pycodestyle==2.11.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0